chore: update allocator

+++ src/kv/storage/cache.rs
@@ -1,15 +1,15 @@
-use crate::kv::storage::inner::{page_s, PAGE_SIZE, gpid_t};
+use crate::kv::storage::inner::{gpid_t, page_s, PAGE_SIZE};
 
 // 1MB for test
-const MAX_CACHE_SIZE: u64 = 1 << 20;
-const MAX_MAPPED_PG: u64 = MAX_CACHE_SIZE / PAGE_SIZE;
+const MAX_CACHE_SIZE: usize = 1 << 20;
+const MAX_MAPPED_PG: usize = MAX_CACHE_SIZE / PAGE_SIZE;
 const EVECT_NUM: i32 = 128;
 
-const PAGE_HASH_NUM: u64 = MAX_MAPPED_PG;
-const PAGE_HASH_MASK: u64 = MAX_MAPPED_PG - 1;
+const PAGE_HASH_NUM: usize = MAX_MAPPED_PG;
+const PAGE_HASH_MASK: usize = MAX_MAPPED_PG - 1;
 
-const PG_DIRTY: u64 = 1 << 0;
-const PG_BUSY: u64 = 1 << 1;
+const PG_DIRTY: usize = 1 << 0;
+const PG_BUSY: usize = 1 << 1;
 
 struct node_s {
     prev: Box<node_s>,
@@ -31,9 +31,9 @@ struct pg_s {
 }
 
 pub struct cache_s {
-    mapped_num: u64,
-    busy_num: u64,
-    free_num: u64,
+    mapped_num: usize,
+    busy_num: usize,
+    free_num: usize,
     hash: [node_s; PAGE_HASH_NUM as usize],
     // free list head
     free: node_s,
